lab2/requirements.txt
fastapi==0.95.2
python-multipart==0.0.6
pydantic==1.10.2
python-jose==3.3.0
uvicorn==0.22.0
